Added a data container for offsets in files
The actual data is a number of (possibly large) files, and for each a
set of offsets within each file. This might be used to train a model to
find strings of text. In this case the offsets would be the start and
end points of the strings to extract.